Fix version crash when connecting to gotosocial
While gotosocial has a PR open to send a Mastodon-compatible version in
the `instance` endpoint, whatever it sends is unlikely to be helpful for
the purpose of version-gating features in the client library.

You can disable version-gating, but it still fails on instantiating
due to GtS's version not matching the semver regex. Fix this by moving
the version_check_mode above it, and not parsing the version if it's
turned off.